+++ huawei_dorado.py
@@ -737,6 +737,7 @@ def _logout(args, session):
         insecure=args.INSECURE,
         method='DELETE',
         no_proxy=args.NO_PROXY,
+        proxy=getattr(args, 'PROXY', None),
         timeout=args.TIMEOUT,
     )
 
@@ -851,6 +852,7 @@ def get_creds(args, force_relogin=False):
             header=header,
             insecure=args.INSECURE,
             no_proxy=args.NO_PROXY,
+            proxy=getattr(args, 'PROXY', None),
             timeout=args.TIMEOUT,
         )
     )
@@ -1044,6 +1046,7 @@ def get_data(endpoint, args, max_attempts=3):
                 header=header,
                 insecure=args.INSECURE,
                 no_proxy=args.NO_PROXY,
+                proxy=getattr(args, 'PROXY', None),
                 response_on_error=True,
                 timeout=args.TIMEOUT,
             )