+++ jitsi.py
@@ -60,6 +60,7 @@ def get_data(args, _type='json'):
         header=headers,
         insecure=args.INSECURE,
         no_proxy=args.NO_PROXY,
+        proxy=getattr(args, 'PROXY', None),
         timeout=args.TIMEOUT,
     )